Struct rusoto_acm_pca::AcmPcaClient

pub struct AcmPcaClient { /* fields omitted */ }

A client for the ACM-PCA API.

Methods

impl AcmPcaClient

pub fn new(region: Region) -> AcmPcaClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> AcmPcaClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,